Type refactoring on BasePagerAdapter, BucketTabsFragment refactorign
There is no need to say adapter type of fragment, instead, FragmentItem could be extended
to provide more data to underlying fragments. So swap Fragment type with FragmentItem.

Also, rename Type to BucketType and move it presenter as view is no good place for data holder class.

BucketTabsFragment refactored to respond to data from presenter, not pull it.